# contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: TI-Toolbox job server (tit.server)
  version: "1.0.0"
  description: >-
    v1 contract for the v3 Electron GUI. Every path and schema from
    `contracts/openapi.v0.yaml` is carried unchanged (the real server's `--dump-openapi` output
    must be a superset of both; CI gate `dev/contracts_check.py`); v1 adds the job engine, catalog
    CRUD, ViewSpec viewers, jailed file access, validate/plan preflight, settings, and the
    `/ws/jobs` live-job stream. Auth is unchanged from v0: the launcher passes
    `?token=<TIT_SERVER_TOKEN>` exactly once; the server mints a random session id (never the
    token), keeps it in memory, sets it as an HttpOnly, SameSite=Strict `tit_session` cookie and
    redirects to `/`. A server restart forgets every session (401 -> back to the launcher). API
    routes accept either that cookie or `Authorization: Bearer <token>` (scripts). Unauthenticated
    requests to any route other than `/api/health` and `/auth/session` get 401.
    `POST /auth/logout` forgets the session and clears the cookie.

    Config request bodies (`config` fields below) are `dict[str, Any]`, validated server-side with
    `tit.config_io.deserialize_config`. Their real shapes are the `contracts/schema.json` `$defs`
    (`SimulationConfig`, `Montage`, `FlexConfig`, `ExConfig`, `MExConfig`, `AnalyzerConfig`,
    `PreprocessConfig`, `QSIPrepConfig`, `QSIReconConfig`, `GroupComparisonConfig`,
...
    today.

...

...

...
    difference is never reported as drift.
...
servers:
...
paths:
...
  /api/catalog/simulations/{name}:
    get:
      tags: [catalog]
      summary: One simulation's full detail (same per-item shape as the list, fetched directly)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: query
          name: subject
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject or simulation
...
  /api/catalog/subjects/{id}:
    get:
      tags: [catalog]
      summary: One subject's full detail (flags, m2m path, eeg nets, leadfields, dwi/ct presence)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/catalog/electrode-overlays:
    get:
      tags: [catalog]
      summary: Electrode-overlay NIfTI presence for one simulation, per TI/mTI mode
      description: >-
        Listing-only: whether `tit.tools.electrode_overlay` (a `tools` job) has already produced
        `<simulation>/<mode>/montage_imgs/electrode_overlay_subject.nii.gz` for each mode, and its
        path. Building the overlay is out of scope here; see `tit/server/routes/viewers.py` and
        `tit/catalog.py::electrode_overlays`.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
        - in: query
          name: simulation
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ElectrodeOverlay" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject or simulation
  /api/catalog/montages:
    get:
      tags: [catalog]
      summary: All montages, grouped by EEG net and polarity kind
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Montages"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/catalog/montages/{net}/{kind}/{name}:
    put:
      tags: [catalog]
      summary: Create or overwrite one montage
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: net
          required: true
          schema: { type: string }
        - in: path
          name: kind
          required: true
          schema: { type: string, enum: [uni_polar, multi_polar] }
        - in: path
          name: name
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pairs]
              properties:
                pairs:
                  $ref: "#/components/schemas/MontagePairs"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MontagePairs"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [catalog]
      summary: Delete one montage
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: net
          required: true
          schema: { type: string }
        - in: path
          name: kind
          required: true
          schema: { type: string, enum: [uni_polar, multi_polar] }
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        "204":
          description: deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/catalog/eeg-nets:
    get:
      tags: [catalog]
      summary: EEG nets available to a subject (electrode labels per net)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EegNet" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject
  /api/catalog/atlases:
    get:
      tags: [catalog]
      summary: Atlases available to a subject in a given space
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
        - in: query
          name: space
          required: false
          schema: { type: string, enum: [subject, mni] }
        - in: query
          name: kind
          required: false
          schema: { type: string, enum: [cortical, subcortical] }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Atlas" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject
  /api/catalog/atlases/regions:
    get:
      tags: [catalog]
      summary: Regions of one atlas, optionally filtered by hemisphere
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
        - in: query
          name: atlas
          required: true
          schema: { type: string }
        - in: query
          name: hemi
          required: false
          schema: { type: string, enum: [lh, rh, both] }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Region" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject or atlas
  /api/catalog/rois:
    get:
      tags: [catalog]
      summary: Saved ROIs for a subject
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Roi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject
    post:
      tags: [catalog]
      summary: Save a new ROI for a subject
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Roi"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roi"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/catalog/rois/{name}:
    delete:
      tags: [catalog]
      summary: Delete one saved ROI
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: query
          name: subject
          required: true
          schema: { type: string }
      responses:
        "204":
          description: deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/catalog/leadfields:
    get:
      tags: [catalog]
      summary: Precomputed leadfields for a subject
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Leadfield" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject
  /api/catalog/flex-runs:
    get:
      tags: [catalog]
      summary: Flex-search runs for a subject
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FlexRun" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject
  /api/catalog/ex-runs:
    get:
      tags: [catalog]
      summary: Ex-search (or mEx-search) runs for a subject
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
        - in: query
          name: kind
          required: false
          schema: { type: string, enum: [ex, mex] }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ExRun" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject
  /api/catalog/ex-runs/{run}/results:
    get:
      tags: [catalog]
      summary: Full results table (every montage/current-split scored) of one ex/mEx run
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: run
          required: true
          schema: { type: string }
        - in: query
          name: subject
          required: true
          schema: { type: string }
        - in: query
          name: kind
          required: true
          schema: { type: string, enum: [ex, mex] }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/catalog/analyses:
    get:
      tags: [catalog]
      summary: Analyzer runs for one subject/simulation
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
        - in: query
          name: simulation
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Analysis" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject or simulation
  /api/catalog/analyses/{name}/summary:
    get:
      tags: [catalog]
      summary: Summary statistics table of one analysis
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: query
          name: subject
          required: true
          schema: { type: string }
        - in: query
          name: simulation
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/catalog/reports:
    get:
      tags: [catalog]
      summary: Generated HTML reports for a subject
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Report" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject
  /api/catalog/freehand:
    get:
      tags: [catalog]
      summary: Saved free-hand electrode configurations for a subject
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: subject
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FreehandConfig" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: unknown subject
  /api/catalog/freehand/{name}:
    put:
      tags: [catalog]
      summary: Create or overwrite one free-hand electrode configuration
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: query
          name: subject
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreehandConfig"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreehandConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/catalog/group:
    get:
      tags: [catalog]
      summary: Project-level group catalog (group stats runs, nilearn visuals, group analyses)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupCatalog"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/catalog/notes:
    get:
      tags: [catalog]
      summary: Quick Notes content
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notes"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [catalog]
      summary: Replace Quick Notes content
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Notes"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notes"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/catalog/subject-info:
    get:
      tags: [catalog]
      summary: Presence matrix (subject x data-stage) across the whole project
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableData"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tit_session
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: missing or invalid session/token
    NotFound:
      description: no such resource
...
  schemas:
...
    SubjectDetail:
      description: v0 Subject plus m2m path, eeg nets, leadfield presence, dwi/ct presence
      allOf:
        - $ref: "#/components/schemas/Subject"
        - type: object
          required: [m2m_path, eeg_nets, has_leadfields, has_dwi, has_ct]
          properties:
            m2m_path: { type: [string, "null"] }
            eeg_nets:
              type: array
              items: { type: string }
            has_leadfields:
              type: array
              items: { type: string }
              description: EEG net names with a precomputed leadfield
            has_dwi: { type: boolean }
            has_ct: { type: boolean }
...
    SimulationDetail:
      description: v0 Simulation plus montages, fields, spaces, reports, niftis, meshes
      allOf:
        - $ref: "#/components/schemas/Simulation"
        - type: object
          required: [montages, fields, space, report_ids, niftis, meshes]
          properties:
            montages:
              type: array
              items: { type: string }
            fields:
              type: array
              items: { type: string }
            space:
              type: array
              items: { type: string, enum: [subject, mni] }
            report_ids:
              type: array
              items: { type: string }
            niftis:
              type: array
              items: { $ref: "#/components/schemas/NiftiRef" }
            meshes:
              type: array
              items: { $ref: "#/components/schemas/MeshRef" }
    MontagePairs:
      type: array
      description: pairs of electrode labels, e.g. [["E1","E2"],["E3","E4"]]
      items:
        type: array
        items: { type: string }
        minItems: 2
        maxItems: 2
...
    Montages:
      type: object
      required: [nets]
      properties:
        nets:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/MontageNet"
    EegNet:
      type: object
      required: [name, electrodes, n]
      properties:
        name: { type: string }
        electrodes:
          type: array
          items: { type: string }
        n: { type: integer }
    Atlas:
      type: object
      required: [id, name, path]
      properties:
        id: { type: string }
        name: { type: string }
        path: { type: string }
        hemispheres:
          type: array
          items: { type: string, enum: [lh, rh] }
    Region:
      type: object
      required: [id, name, hemi]
      description: >-
        For a cortical (surface/annotation) atlas, `id` is the FreeSurfer `.annot` label index --
        exactly the integer `FlexConfig.AtlasROI.label` / `ExConfig`'s equivalent needs, not a
        display string -- and `hemi` is that region's own hemisphere ("lh"/"rh"), since one
        `.annot` file is per-hemisphere and a cortical ROI can union regions across both. For a
        subcortical (volumetric) atlas, `id` is the voxel label value in that atlas's NIfTI and
        `hemi` is `null` (volumetric atlases are not inherently split by hemisphere).
      properties:
        id: { type: integer }
        name: { type: string }
        hemi:
          type: [string, "null"]
          enum: [lh, rh, null]
    Roi:
      type: object
      required: [name, x, y, z, space]
      properties:
        name: { type: string }
        x: { type: number }
        y: { type: number }
        z: { type: number }
        radius: { type: number }
        space: { type: string, enum: [subject, mni] }
    Leadfield:
      type: object
      required: [net, path, exists, size_bytes]
      properties:
        net: { type: string }
        path: { type: string }
        exists: { type: boolean }
        size_bytes: { type: integer }
    FlexRun:
      type: object
      required: [name, path, goal, roi, created, manifest, artifacts]
      properties:
        name: { type: string }
        path: { type: string }
        goal: { type: string }
        roi:
          type: object
          additionalProperties: true
        created: { type: string, format: date-time }
        manifest:
          type: object
          additionalProperties: true
        artifacts:
          type: array
          items: { $ref: "#/components/schemas/Artifact" }
          description: run outputs (Pareto-front/convergence PNGs, CSVs, manifests) to list, not just reveal-in-folder
...
    ExRun:
      type: object
      required: [run_name, path, eeg_net, created, artifacts]
      properties:
        run_name: { type: string }
        path: { type: string }
        eeg_net: { type: string }
        created: { type: string, format: date-time }
        best:
          oneOf:
            - $ref: "#/components/schemas/ExBest"
            - type: "null"
        artifacts:
          type: array
          items: { $ref: "#/components/schemas/Artifact" }
          description: run outputs (PNG plots, results CSV, config JSON) to list on the Results/Optimizer pages
    Analysis:
      type: object
      required: [name, space, field, roi, csv, json]
      properties:
        name: { type: string }
        space: { type: string, enum: [subject, mni] }
        field: { type: string }
        roi: { type: string }
        csv: { type: string }
        json: { type: string }
        msh: { type: [string, "null"] }
        nifti: { type: [string, "null"] }
        pdf: { type: [string, "null"] }
    Report:
      type: object
      required: [id, kind, title, path, created]
      properties:
        id: { type: string }
        kind: { type: string }
        title: { type: string }
        path: { type: string }
        created: { type: string, format: date-time }
    ElectrodeOverlay:
      type: object
      required: [mode, path, exists]
      properties:
        mode: { type: string, enum: [TI, mTI] }
        path: { type: string, description: container path (may not exist yet -- see `exists`) }
        exists: { type: boolean }
...
    FreehandConfig:
      type: object
      required: [name, type, electrode_positions]
      properties:
        name: { type: string }
        type:
          type: string
          enum: [U, M]
          description: >-
            On-disk stim_configs semantics (unipolar/multipolar), not xyz/label -- see
            `tit/catalog.py::_read_freehand_file`, the source of truth for
            `m2m_<id>/stim_configs/*.json` written by
            `tit/gui/extensions/electrode_placement.py`.
        electrode_positions:
          type: array
          items: { $ref: "#/components/schemas/ElectrodePosition" }
...
    TableData:
      type: object
      required: [columns, rows]
      properties:
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          items:
            type: array
            items: {}

    # ------------------------------------------------------------------------------- overview (v1)
...